+++ Fuji/Source/MFTexture_Internal.h
@@ -90,11 +90,11 @@ public:
 };
 #endif
 
-extern uint32 gMFImagePlatformFormat[MFRD_Max][ImgFmt_Max];
+extern uint8 gMFImagePlatformFormat[MFRD_Max][ImgFmt_Max];
 
 inline uint32 MFTexture_GetPlatformFormatID(MFImageFormat format, MFRendererDrivers rendererDriver)
 {
-	return gMFImagePlatformFormat[rendererDriver][format];
+	return (uint32)gMFImagePlatformFormat[rendererDriver][format];
 }
 
 #endif
